add method for fetch photo
