+++ ChomboGBFinal/Core/Network/Firebase/CloudFirestore.swift
@@ -71,36 +71,6 @@ final class CloudFirestore {
         }
     }
     
-    func loadImage(inputImage: UIImage) {
-        
-        let randomID = UUID.init().uuidString
-        let uploadRef = Storage.storage().reference(withPath: "toolsImage/\(randomID).jpg")
-        guard let imageData = inputImage.jpegData(compressionQuality: 0.75) else { return }
-        let uploadMetadata = StorageMetadata.init()
-        uploadMetadata.contentType = "image/jpeg"
-        
-        uploadRef.putData(imageData, metadata: uploadMetadata) { (downloadMetadata, error) in
-            if let error = error {
-                print("Возникла ошибка: \(error.localizedDescription)")
-                return
-            }
-            print("Загрузка выполнена успешно: \(String(describing: downloadMetadata))")
-        }
-    }
-    
-//    func fetchImage() {
-//        let storageRef = Storage.storage().reference(withPath: "toolsImage/testPicture.jpg")
-//        storageRef.getData(maxSize: 4 * 1024 * 1024) { [weak self] (data, error) in
-//            if let error = error {
-//                print("Ошибка при получении данных: \(error.localizedDescription)")
-//                return
-//            }
-//            if let data = data {
-//                self?.imageView.image = UIImage(data: data)
-//            }
-//        }
-//    }
-    
 }
